pyproject.toml
[project]
name = "mm-shortvideo-rec"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gdown>=6.0.0",
    "kaggle>=2.0.2",
    "kagglehub>=1.0.0",
    "pandas>=3.0.2",
    "pillow>=12.1.1",
    "streamlit>=1.57.0",
    "streamlit-shadcn-ui>=0.1.19",
    "tensorboard>=2.20.0",
    "tensorboardx>=2.6.5",
    "torch>=2.11.0",
    "torchvision>=0.26.0",
    "tqdm>=4.66.5",
    "transformers>=5.5.4",
]

[[tool.uv.index]]
url = "https://download.pytorch.org/whl/cu126"